requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-bmp3xx==1.3.20
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-scd4x==1.4.5
adafruit-circuitpython-sht4x==1.0.21
adafruit-circuitpython-tca9548a==0.7.4
adafruit-circuitpython-typing==1.11.2
adafruit-circuitpython-veml7700==2.0.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.1
anyio==4.7.0
anytree==2.12.1
APScheduler==3.11.0
binho-host-adapter==0.1.6
click==8.1.7
h11==0.14.0
idna==3.10
lgpio==0.2.2.0
prometheus_client==0.21.1
pyftdi==0.56.0
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0.2
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sysv-ipc==1.1.0
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.1
